requirements.txt
# ============================================
# PhoneAgent - Python 依赖
# ============================================
# 手动安装: bash scripts/install/install_server.sh
# ============================================

# AI 模型
openai>=1.0.0

# HTTP 客户端 (AI 模型调用 + PC Agent 通信)
httpx>=0.27.0

# 图像处理
Pillow>=10.0.0

# 数值计算 (PC Agent 可访问性树过滤)
numpy>=1.24.0

# Web 框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# 数据验证和配置
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# 数据库
sqlalchemy>=2.0.0

# 音频处理
pydub>=0.25.1

# ============================================
# 可选依赖 (性能优化)
# ============================================
# OCR (手机端文本识别优化)
# 取消注释以下两行以启用 OCR:
# paddlepaddle>=2.5.0
# paddleocr>=2.7.0